fix scroll area and add the notes to the search dialog
