+++ ink-down-web/src/app/(private)/components/notes/tags-list.tsx
@@ -1,10 +1,10 @@
 "use client"
+
 import type { TagsDataType } from "@/app/@types/note-types";
 import { SidebarMenuItem, SidebarMenuButton } from "@/components/ui/sidebar";
 import { useTagsActions } from "../../../../hooks/use-tags";
-import { useCallback } from "react";
+import { useCallback, useEffect } from "react";
 import { TagContext } from "./tag-context";
-import { ScrollArea } from "@/components/ui/scroll-area";
 
 export const TagList = ({ initialTags }: { initialTags: TagsDataType[] }) => {
   const {
@@ -13,7 +13,7 @@ export const TagList = ({ initialTags }: { initialTags: TagsDataType[] }) => {
     handleChangeColor,
     handleDeleteTag
   } = useTagsActions(initialTags);
-
+  
   // Callback que previne rerender da parte do react.
   const memoizedRename = useCallback(
     (id: number, newName: string) => handleRenameTag(id, newName),
@@ -21,27 +21,29 @@ export const TagList = ({ initialTags }: { initialTags: TagsDataType[] }) => {
   );
 
   return (
-    <ScrollArea className="h-[800px]">
-      {tags.map((tag) => (
-        <div key={tag.id} className="context-menu-no-close">
-          <TagContext
-            tag={tag}
-            onRename={memoizedRename}
-            onChangeColor={handleChangeColor}
-            onDelete={handleDeleteTag}
-          >
-            <SidebarMenuItem className="list-none">
-              <SidebarMenuButton>
-                <div
-                  style={{ backgroundColor: tag.colorHex }}
-                  className="w-5 h-5 border-muted-foreground rounded-md border-[1px]"
-                />
-                {tag.name}
-              </SidebarMenuButton>
-            </SidebarMenuItem>
-          </TagContext>
-        </div>
-      ))}
-    </ScrollArea>
+    <div className="max-h-[60vh] overflow-hidden">
+        {tags.map((tag) => (
+          <div key={tag.id} className="context-menu-no-close">
+            <TagContext
+              tag={tag}
+              onRename={memoizedRename}
+              onChangeColor={handleChangeColor}
+              onDelete={handleDeleteTag}
+            >
+              <SidebarMenuItem className="list-none">
+                <SidebarMenuButton>
+                  <div
+                    style={{ backgroundColor: tag.colorHex }}
+                    className="w-5 h-5 border-muted-foreground rounded-md border-[1px]"
+                  />
+                  <span className="truncate">
+                    {tag.name}
+                  </span>
+                </SidebarMenuButton>
+              </SidebarMenuItem>
+            </TagContext>
+          </div>
+        ))}
+    </div>
   );
 };
