+++ ink-down-web/src/app/(private)/settings/page.tsx
@@ -1,0 +1,7 @@
+export default function SettingsPage(){
+  return (
+    <h1>
+      Settings
+    </h1>
+  )
+}
